refactor: tagInput 상태관리 hook 구현
